pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "driveforge"
version = "0.1.0"
description = "In-house enterprise drive refurbishment pipeline"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "John Thompson", email = "jthompson4862@gmail.com" }]
keywords = ["drive", "refurbish", "smart", "badblocks", "homelab"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Environment :: Web Environment",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: System :: Hardware",
  "Topic :: System :: Systems Administration",
]

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "jinja2>=3.1",
  "httpx>=0.27",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "click>=8.1",
  "pyyaml>=6.0",
  "qrcode[pil]>=7.4",
  "pillow>=10.2",
  "rich>=13.7",
  "textual>=0.60",
  "brother-ql>=0.9.4",
]

[project.optional-dependencies]
linux = [
  "pyudev>=0.24",
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "ruff>=0.3",
  "mypy>=1.9",
  "httpx>=0.27",
]

[project.scripts]
driveforge = "driveforge.cli:main"
driveforge-daemon = "driveforge.daemon.app:main"
driveforge-tui = "driveforge.tui.app:main"

[project.urls]
Homepage = "https://github.com/JT4862/driveforge"
Repository = "https://github.com/JT4862/driveforge"
Issues = "https://github.com/JT4862/driveforge/issues"

[tool.hatch.build.targets.wheel]
packages = ["driveforge"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
